Safari-611.1.10/LayoutTests/media/track/track-delete-during-setup-expected.txt

>

EXPECTED (track1.readyState == '0') OK
EXPECTED (track1.track.mode == 'disabled') OK

END OF TEST
